Se renombran varios preprocesos de imagen Se terminan de añadir las propiedades de los diferentes metodos de preprocesamiento y de detección /descripción de puntos de interés en la configuración del programa

+++ src/fme/widgets/HmclaheWidget.cpp
@@ -66,13 +66,13 @@ void HmclaheWidget::setBlockSize(const QSize &blockSize)
 void HmclaheWidget::setL(double l)
 {
   const QSignalBlocker blockerL(mL);
-  mL->setValue(l);
+  mL->setValue(static_cast<double>(l));
 }
 
 void HmclaheWidget::setPhi(double phi)
 {
   const QSignalBlocker blockerPhi(mPhi);
-  mPhi->setValue(phi);
+  mPhi->setValue(static_cast<double>(phi));
 }
 
 void HmclaheWidget::update()
